Prepare mockito's codegen tests for changes to include inherited members.
Many of these codegen tests used an entire library content as their expectation. Sometimes the assertion was that a certain method or constructor _wasn't present_, which was conveyed indirectly, and vaguely.

When classes will start to override inherited methods and fields, like hashCode and toString(), the generated mock classes will get bigger, and it is no longer viable to write tests like this.

Since this is a huge refactoring of these tests, I'm landing this preparation change first.

https://github.com/dart-lang/mockito/issues/297

PiperOrigin-RevId: 343529871